[#3] Add a button to close open topics
Clicking the Close button also clears the current URL fragment, adding
to history.  This permits the immediate reopening of the closed topic:
if its title remained in the hash, clicking the link would not trigger
OnLocationChange.  The code used to clear the URL fragment is:

    Navigation.newUrl "#"

This is mildly hacky and exploits several subtle behaviors.  First,
this doesn't actually remove the URL fragment, but replaces it with an
empty fragment.  The only way to actually remove the URL fragment is
to provide the current page's URL sans-fragment, which is inconvenient
to obtain in `update`.  Second, empty URL fragments cannot refer to
elements with empty id attributes, but instead cause most browsers to
scroll to the top of the page.  `newUrl`, however, does not appear to
affect the scroll position of the page at all.  Lastly, `newUrl` can
and does trigger OnLocationChange, but in this case `location.hash`
remains completely empty (as opposed to containing "#" by itself), so
no call to `getTopic` is made.